pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "promptframe"
version = "0.1.0"

authors = [
    {name = "Anik Mallick", email = "anik.mallick.code@gmail.com"}
]

description = "LLM-agnostic prompt management — YAML-based storage, typed loading, and a fluent builder API."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["llm", "prompt", "prompt-engineering", "yaml"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    # "pyyaml>=5.1",
    "rich>=13.0.0",
    "python-frontmatter>=1.1.0",
]

[project.optional-dependencies]
ui = ["streamlit>=1.30"]
dev = ["pytest>=7", "ruff", "mypy", "build", "twine"]

[project.scripts]
promptframe = "promptframe.cli:main"

[project.urls]
Repository = "https://github.com/AnikMallick/promptframe"

# [tool.setuptools.packages.find]
# where = ["."]
# include = ["promptframe*"]
[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true
